requirements.txt
appdirs==1.4.4
asgiref==3.5.2
autopep8
black==21.6b0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.7
django-autoslug==1.9.8
django-countries==7.2.1
django-environ==0.7.0
django-extensions==3.1.5
django-filter==21.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environ==1.0
flake8==3.9.2
idna==3.3
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
phonenumbers==8.12.33
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2022.1
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
